Now using indent -linux -nut

+++ src/Cache_inode/cache_inode_link.c
@@ -90,7 +90,7 @@
 
 #ifdef _SOLARIS
 #include "solaris_port.h"
-#endif				/* _SOLARIS */
+#endif                          /* _SOLARIS */
 
 #include "LRU_List.h"
 #include "log_functions.h"
@@ -127,13 +127,13 @@
  *
  */
 cache_inode_status_t cache_inode_link(cache_entry_t * pentry_src,
-				      cache_entry_t * pentry_dir_dest,
-				      fsal_name_t * plink_name,
-				      fsal_attrib_list_t * pattr,
-				      hash_table_t * ht,
-				      cache_inode_client_t * pclient,
-				      fsal_op_context_t * pcontext,
-				      cache_inode_status_t * pstatus)
+                                      cache_entry_t * pentry_dir_dest,
+                                      fsal_name_t * plink_name,
+                                      fsal_attrib_list_t * pattr,
+                                      hash_table_t * ht,
+                                      cache_inode_client_t * pclient,
+                                      fsal_op_context_t * pcontext,
+                                      cache_inode_status_t * pstatus)
 {
   fsal_status_t fsal_status;
   fsal_handle_t handle_src;
@@ -170,9 +170,9 @@ cache_inode_status_t cache_inode_link(cache_entry_t * pentry_src,
 
   /* Check if an entry of the same name doesn't exist in the destination directory */
   if ((pentry_lookup = cache_inode_lookup(pentry_dir_dest,
-					  plink_name,
-					  &lookup_attributes,
-					  ht, pclient, pcontext, pstatus)) != NULL)
+                                          plink_name,
+                                          &lookup_attributes,
+                                          ht, pclient, pcontext, pstatus)) != NULL)
     {
       /* There exists such an entry... */
       *pstatus = CACHE_INODE_ENTRY_EXISTS;
@@ -233,8 +233,8 @@ cache_inode_status_t cache_inode_link(cache_entry_t * pentry_src,
 
     default:
       DisplayLogJdLevel(pclient->log_outputs, NIV_CRIT,
-			"WARNING: unknown source pentry type: internal_md.type=%d, line %d in file %s",
-			pentry_src->internal_md.type, __LINE__, __FILE__);
+                        "WARNING: unknown source pentry type: internal_md.type=%d, line %d in file %s",
+                        pentry_src->internal_md.type, __LINE__, __FILE__);
       *pstatus = CACHE_INODE_BAD_TYPE;
       pclient->stat.func_stats.nb_err_unrecover[CACHE_INODE_LINK] += 1;
       return *pstatus;
@@ -271,7 +271,7 @@ cache_inode_status_t cache_inode_link(cache_entry_t * pentry_src,
   cache_inode_get_attributes(pentry_dir_dest, &dirdest_attributes);
   fsal_status =
       MFSL_link(&pentry_src->mobject, &pentry_dir_dest->mobject, plink_name, pcontext,
-		&pclient->mfsl_context, &link_attributes, &dirdest_attributes);
+                &pclient->mfsl_context, &link_attributes, &dirdest_attributes);
 #else
   fsal_status =
       FSAL_link(&handle_src, &handle_dest, plink_name, pcontext, &link_attributes);
@@ -283,42 +283,42 @@ cache_inode_status_t cache_inode_link(cache_entry_t * pentry_src,
       V_w(&pentry_src->lock);
 
       if (fsal_status.major == ERR_FSAL_STALE)
-	{
-	  cache_inode_status_t kill_status;
-	  fsal_status_t getattr_status;
+        {
+          cache_inode_status_t kill_status;
+          fsal_status_t getattr_status;
 
-	  DisplayLog
-	      ("cache_inode_link: Stale FSAL File Handle detected for at least one in  pentry = %p and pentry = %p",
-	       pentry_src, pentry_dir_dest);
+          DisplayLog
+              ("cache_inode_link: Stale FSAL File Handle detected for at least one in  pentry = %p and pentry = %p",
+               pentry_src, pentry_dir_dest);
 
-	  /* Use FSAL_getattrs to find which entry is staled */
-	  getattr_status = FSAL_getattrs(&handle_src, pcontext, &link_attributes);
-	  if (getattr_status.major == ERR_FSAL_ACCESS)
-	    {
-	      DisplayLog
-		  ("cache_inode_link: Stale FSAL File Handle detected for pentry = %p",
-		   pentry_src);
+          /* Use FSAL_getattrs to find which entry is staled */
+          getattr_status = FSAL_getattrs(&handle_src, pcontext, &link_attributes);
+          if (getattr_status.major == ERR_FSAL_ACCESS)
+            {
+              DisplayLog
+                  ("cache_inode_link: Stale FSAL File Handle detected for pentry = %p",
+                   pentry_src);
 
-	      if (cache_inode_kill_entry(pentry_src, ht, pclient, &kill_status) !=
-		  CACHE_INODE_SUCCESS)
-		DisplayLog("cache_inode_link: Could not kill entry %p, status = %u",
-			   pentry_src, kill_status);
-	    }
+              if (cache_inode_kill_entry(pentry_src, ht, pclient, &kill_status) !=
+                  CACHE_INODE_SUCCESS)
+                DisplayLog("cache_inode_link: Could not kill entry %p, status = %u",
+                           pentry_src, kill_status);
+            }
 
-	  getattr_status = FSAL_getattrs(&handle_dest, pcontext, &link_attributes);
-	  if (getattr_status.major == ERR_FSAL_ACCESS)
-	    {
-	      DisplayLog
-		  ("cache_inode_link: Stale FSAL File Handle detected for pentry = %p",
-		   pentry_dir_dest);
+          getattr_status = FSAL_getattrs(&handle_dest, pcontext, &link_attributes);
+          if (getattr_status.major == ERR_FSAL_ACCESS)
+            {
+              DisplayLog
+                  ("cache_inode_link: Stale FSAL File Handle detected for pentry = %p",
+                   pentry_dir_dest);
 
-	      if (cache_inode_kill_entry(pentry_dir_dest, ht, pclient, &kill_status) !=
-		  CACHE_INODE_SUCCESS)
-		DisplayLog("cache_inode_link: Could not kill entry %p, status = %u",
-			   pentry_dir_dest, kill_status);
-	    }
+              if (cache_inode_kill_entry(pentry_dir_dest, ht, pclient, &kill_status) !=
+                  CACHE_INODE_SUCCESS)
+                DisplayLog("cache_inode_link: Could not kill entry %p, status = %u",
+                           pentry_dir_dest, kill_status);
+            }
 
-	}
+        }
       *pstatus = CACHE_INODE_FSAL_ESTALE;
 
       return *pstatus;
@@ -353,18 +353,18 @@ cache_inode_status_t cache_inode_link(cache_entry_t * pentry_src,
 
     default:
       DisplayLogJdLevel(pclient->log_outputs, NIV_CRIT,
-			"WARNING: Major type incoherency line %d in file %s", __LINE__,
-			__FILE__);
+                        "WARNING: Major type incoherency line %d in file %s", __LINE__,
+                        __FILE__);
       break;
     }
 
   /* Add the new entry in the destination directory */
   if (cache_inode_add_cached_dirent(pentry_dir_dest,
-				    plink_name,
-				    pentry_src,
-				    NULL,
-				    ht,
-				    pclient, pcontext, &status) != CACHE_INODE_SUCCESS)
+                                    plink_name,
+                                    pentry_src,
+                                    NULL,
+                                    ht,
+                                    pclient, pcontext, &status) != CACHE_INODE_SUCCESS)
     {
       V_w(&pentry_dir_dest->lock);
       V_w(&pentry_src->lock);
